+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- New middleware `responseHeadersOnStatus` ([GH-84](https://github.com/uniport/uniport-gateway/issues/84)).
 - CorsMiddleware deduplicates the headers ([GH-85](https://github.com/uniport/uniport-gateway/issues/85)).
 
 ## [10.3.0]-202602121511-92-c7ee225 - 2026-02-12